Build the storage mangager
The storage manager is the front facing object
that manage volumes, disks and devices

This will be the main object to be exposed by the storage
api that should come later
